chore: move AdminUser
