+++ src/Security/AdminUserProvider.php
@@ -2,7 +2,6 @@
 
 namespace App\Security;
 
-use App\Entity\AdminUser;
 use Symfony\Component\Security\Core\Exception\UnsupportedUserException;
 use Symfony\Component\Security\Core\Exception\UsernameNotFoundException;
 use Symfony\Component\Security\Core\User\UserInterface;
